+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased][unreleased]
+### Added
+- all Parameters for alias-usage
+- allowing to pass an array of additional parameters to the `PostFinanceEPayment->createPayment` method
+
 ## 0.1.0 Alpha 1
 ### Added
 - handle payment request